+++ gaps/shower/src/colours.cu
@@ -2,7 +2,7 @@
 
 __device__ void MakeColours(Event &ev, int *coli, int *colj, const int flavs[3],
                             const int colij[2], const int colk[2],
-                            const int rand) {
+                            const double rand) {
   // Increase variable ev.GetShowerC() by 1
   ev.IncrementShowerC();
